+++ libs/newsletter-workflow/src/lib/newsletter-workflow.ts
@@ -1,9 +1,9 @@
 import type { WizardLayout } from '@newsletters-nx/state-machine';
 import { cancelLayout } from './steps/cancelLayout';
 import { createNewsletterLayout } from './steps/createNewsletterLayout';
-import { descriptionLayout } from './steps/description';
+import { descriptionLayout } from './steps/descriptionLayout';
 import { finishLayout } from './steps/finishLayout';
-import { pillarLayout } from './steps/pillar';
+import { pillarLayout } from './steps/pillarLayout';
 
 export const newslettersWorkflowStepLayout: WizardLayout = {
 	createNewsletter: createNewsletterLayout,
